CAbstractTextEdit: Fix 'paste' action
If there is a selection then CursorPos matches the mouse press pos,
but after the selection text removed the CursorPos (which is the place
for Clipboard Text insertion) wasn't synced and can point out of the
current TextEdit string and that leads to no op.

Sync CursorPos to the selection beginning point to fix invalid inserts.